+++ tests/adminserver.test.ts
@@ -1,15 +1,15 @@
-import { DBOS, StatusString } from '../../src';
-import { DBOSConfig } from '../../src/dbos-executor';
-import { WorkflowQueue } from '../../src';
-import { generateDBOSTestConfig, queueEntriesAreCleanedUp, setUpDBOSTestDb } from '../helpers';
-import { QueueMetadataResponse } from '../../src/httpServer/server';
-import { HealthUrl, WorkflowQueuesMetadataUrl, WorkflowRecoveryUrl } from '../../src/httpServer/server';
-import { globalParams, sleepms } from '../../src/utils';
+import { DBOS, StatusString } from '../src';
+import { DBOSConfig } from '../src/dbos-executor';
+import { WorkflowQueue } from '../src';
+import { generateDBOSTestConfig, queueEntriesAreCleanedUp, setUpDBOSTestDb } from './helpers';
+import { QueueMetadataResponse } from '../src/httpServer/server';
+import { HealthUrl, WorkflowQueuesMetadataUrl, WorkflowRecoveryUrl } from '../src/httpServer/server';
+import { globalParams, sleepms } from '../src/utils';
 import { Client } from 'pg';
-import { step_info } from '../../schemas/system_db_schema';
+import { step_info } from '../schemas/system_db_schema';
 import http from 'http';
-import { DBOSWorkflowCancelledError } from '../../src/error';
-import * as protocol from '../../src/conductor/protocol';
+import { DBOSWorkflowCancelledError } from '../src/error';
+import * as protocol from '../src/conductor/protocol';
 
 // Add type definitions for admin server API responses
 interface ErrorResponse {
