--- tests/prisma/.env
@@ -1,1 +1,0 @@
-DATABASE_URL="postgresql://postgres:${PGPASSWORD}@localhost:5432/dbostest?schema=public"
