--- tests/prisma/schema.prisma
@@ -1,21 +1,0 @@
-// This is your Prisma schema file,
-// learn more about it in the docs: https://pris.ly/d/prisma-schema
-
-generator client {
-  provider = "prisma-client-js"
-}
-
-datasource db {
-  provider = "postgresql"
-  url      = env("DATABASE_URL")
-}
-
-model testkv {
-  id           String @id
-  value        String
-}
-
-model dbos_test_user {
-  id           Int @id
-  username     String
-}
